+++ CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [2.4.6] - 2020-11-06 - [#101](https://github.com/mza921/Plex-Auto-Collections/pull/101)
 ### Added
-- Progress for filtering movies/shows
+- Progress for filtering movies/shows and imdb lists
 
 ### Fixed
 - [#93](https://github.com/mza921/Plex-Auto-Collections/issues/93) - actually fixed `max_age`